+++ src/main/java/br/senai/sc/engajamento/reacoes/service/ReacaoService.java
@@ -30,36 +30,45 @@ public class ReacaoService {
     public void criar(@Valid CriarReacaoCommand cmd) {
         Usuario usuario = usuarioRepository.getById(cmd.getIdUsuario());
         Video video = videoRepository.getById(cmd.getIdVideo());
-
         Reacao reacaoExistente = repository.findByIdUsuarioAndIdVideo(usuario, video);
 
-        if (reacaoExistente == null) {
-            Reacao reacao = new Reacao();
-            reacao.setIdUsuario(usuario);
-            reacao.setIdVideo(video);
-            reacao.setCurtida(cmd.getCurtida());
+        if (!video.getEhInativado()) {
+            if (reacaoExistente == null) {
+                Reacao reacao = new Reacao();
+                reacao.setIdUsuario(usuario);
+                reacao.setIdVideo(video);
+                reacao.setCurtida(cmd.getCurtida());
 
-            if(cmd.getCurtida()){
-                video.setQtdCurtidas(video.getQtdCurtidas() + 1);
+                if (cmd.getCurtida()) {
+                    video.setQtdCurtidas(video.getQtdCurtidas() + 1);
+                } else {
+                    video.setQtdDescurtidas(video.getQtdDescurtidas() + 1);
+                }
+                videoService.editarPontuacao(video);
+
+                repository.save(reacao);
+            } else if (reacaoExistente.isCurtida() == cmd.getCurtida()) {
+                if (cmd.getCurtida()) {
+                    video.setQtdCurtidas(video.getQtdCurtidas() - 1);
+                } else {
+                    video.setQtdDescurtidas(video.getQtdDescurtidas() - 1);
+                }
+                videoService.editarPontuacao(video);
+
+                repository.deleteByIdUsuarioAndIdVideo(usuario, video);
             } else {
-                video.setQtdDescurtidas(video.getQtdDescurtidas() + 1);
+                reacaoExistente.setCurtida(!reacaoExistente.isCurtida());
+                if (cmd.getCurtida()) {
+                    video.setQtdCurtidas(video.getQtdCurtidas() + 1);
+                    video.setQtdDescurtidas(video.getQtdDescurtidas() - 1);
+                } else {
+                    video.setQtdCurtidas(video.getQtdCurtidas() - 1);
+                    video.setQtdDescurtidas(video.getQtdDescurtidas() + 1);
+                }
+                repository.save(reacaoExistente);
             }
-            videoService.editarPontuacao(video);
-
-            repository.save(reacao);
-        } else if (reacaoExistente.isCurtida() == cmd.getCurtida()) {
-
-            if(cmd.getCurtida()){
-                video.setQtdCurtidas(video.getQtdCurtidas() - 1);
-            } else {
-                video.setQtdDescurtidas(video.getQtdDescurtidas() - 1);
-            }
-            videoService.editarPontuacao(video);
-
-            repository.deleteByIdUsuarioAndIdVideo(usuario, video);
         } else {
-            reacaoExistente.setCurtida(!reacaoExistente.isCurtida());
-            repository.save(reacaoExistente);
+            throw new NaoEncontradoException("Vídeo não encontrado");
         }
     }
 
@@ -67,17 +76,21 @@ public class ReacaoService {
         Usuario usuario = usuarioRepository.getById(cmd.getIdUsuario());
         Video video = videoRepository.getById(cmd.getIdVideo());
 
-        Reacao reacao = repository.findByIdUsuarioAndIdVideo(usuario, video);
-        if (reacao == null) {
-            throw new NaoEncontradoException("Reação não encontrada!");
+        if (!video.getEhInativado()) {
+            Reacao reacao = repository.findByIdUsuarioAndIdVideo(usuario, video);
+            if (reacao == null) {
+                throw new NaoEncontradoException("Reação não encontrada!");
+            }
+            return reacao;
         }
-
-        return reacao;
+        throw new NaoEncontradoException("Vídeo não encontrado");
     }
 
     public List<Reacao> buscarTodosPorVideo(@Valid BuscarTodosPorVideoReacaoCommand cmd) {
         Video video = videoRepository.getById(cmd.getIdVideo());
-
-        return repository.findAllByIdVideo(video);
+        if (!video.getEhInativado()) {
+            return repository.findAllByIdVideo(video);
+        }
+        throw new NaoEncontradoException("Vídeo não encontrado");
     }
 }
